+++ cryptofeed/exceptions.py
@@ -12,3 +12,11 @@ class MissingSequenceNumber(Exception):
 
 class MissingMessage(Exception):
     pass
+
+
+class UnsupportedTradingPair(Exception):
+    pass
+
+
+class UnsupportedDataFeed(Exception):
+    pass