+++ examples/demo_book_delta.py
@@ -8,8 +8,8 @@ from copy import deepcopy
 
 from cryptofeed.callback import BookCallback, BookUpdateCallback
 from cryptofeed import FeedHandler
-from cryptofeed.exchanges import Bitmex, Coinbase, Bitfinex, Gemini, HitBTC, Poloniex
-from cryptofeed.defines import L2_BOOK, L3_BOOK, BID, ASK, UPD, DEL, BOOK_DELTA
+from cryptofeed.exchanges import Bitmex, Coinbase, Bitfinex, Gemini, HitBTC, Poloniex, Kraken, OKCoin
+from cryptofeed.defines import L2_BOOK, L3_BOOK, BID, ASK, BOOK_DELTA
 
 
 BOOK = None
@@ -47,29 +47,22 @@ async def delta(feed, pair, update, timestamp):
     # handle updates for L2 books
     global BOOK
     for side in (BID, ASK):
-        if UPD in update[side]:
-            for price, size in update[side][UPD]:
-                BOOK[side][price] = size
-        if DEL in update[side]:
-            for price in update[side][DEL]:
+        for price, size in update[side]:
+            if size == 0:
                 del BOOK[side][price]
+            else:
+                BOOK[side][price] = size
 
 
 async def l3_delta(feed, pair, update, timestamp):
     global BOOK
     for side in (BID, ASK):
-        """
-        you must DEL before UPD because a modified order will
-        be sent as a DEL on the old order followed by an UPD
-        with the new price.
-        """
-        if DEL in update[side]:
-            for order, price in update[side][DEL]:
+        for order, price, size in update[side]:
+            if size == 0:
                 del BOOK[side][price][order]
                 if len(BOOK[side][price]) == 0:
                     del BOOK[side][price]
-        if UPD in update[side]:
-            for order, price, size in update[side][UPD]:
+            else:
                 if price in BOOK[side]:
                     BOOK[side][price][order] = size
                 else:
@@ -79,14 +72,16 @@ async def l3_delta(feed, pair, update, timestamp):
 def main():
     f = FeedHandler()
     # due to the way the test verification works, you can only run one for the test
-    # f.add_feed(Bitmex(pairs=['XBTUSD'], channels=[L3_BOOK], callbacks={L3_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(l3_delta)}))
+    #f.add_feed(Bitmex(pairs=['XBTUSD'], channels=[L3_BOOK], callbacks={L3_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(l3_delta)}))
     # f.add_feed(Bitfinex(pairs=['BTC-USD'], channels=[L3_BOOK], callbacks={L3_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(l3_delta)}))
     # f.add_feed(Bitfinex(pairs=['BTC-USD'], channels=[L2_BOOK], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
     # f.add_feed(Coinbase(pairs=['BTC-USD'], channels=[L3_BOOK], callbacks={L3_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(l3_delta)}))
     # f.add_feed(Coinbase(pairs=['BTC-USD'], channels=[L2_BOOK], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
     # f.add_feed(Gemini(book_interval=100, pairs=['BTC-USD'], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
     # f.add_feed(HitBTC(book_interval=100, pairs=['BTC-USD'], channels=[L2_BOOK], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
-    f.add_feed(Poloniex(book_interval=100, channels=['BTC-USDT'], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
+    # f.add_feed(Poloniex(book_interval=100, channels=['BTC-USDT'], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
+    # f.add_feed(Kraken(book_interval=100, pairs=['BTC-USD'], channels=[L2_BOOK], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
+    f.add_feed(OKCoin(book_interval=100, pairs=['BTC-USD'], channels=[L2_BOOK], callbacks={L2_BOOK: BookCallback(book), BOOK_DELTA: BookUpdateCallback(delta)}))
 
     f.run()