+++ CHANGES.md
@@ -1,9 +1,22 @@
 ## Changelog
 
-### 0.20.2
-  * Bugfix: Kraken REST api using wrong symbol for trades
-  * Feature: Complete work on standardizing Bitfinex REST API
+### 0.21.1
+  * Feature: Config support for Coinbene, Binance, EXX, BitMEX, Bitfinex, Bitstamp, HitBTC
   * Feature: Complete work on Bitmex REST API
+
+### 0.21.0 (2019-04-07)
+  * Bugfix: Coinbase L3 Book would get in cycle of reconnecting due to missing sequence numbers
+  * Feature: Kraken L2 Book Deltas
+  * Feature: Book deltas streamlined and retain ordering
+  * Feature: OKCoin exchange support
+  * Feature: OKEx exchange support
+  * Feature: Coinbene exchange support
+  * Feature: Support Huobi Global and Huobi USA
+
+### 0.20.2 (2019-03-19)
+  * Bugfix: Kraken REST api using wrong symbol for trades
+  * Feature: Complete work on standardizing Bitfinex rest API
+  * Bugfix: Allow index symbols on Bitmex
 
 ### 0.20.1 (2019-02-16)
   * Feature: Trades sides are now labeled as Buy / Sell instead of Bid / Ask.